requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pywebview==4.1
python-multipart==0.0.5
pydantic==1.8.2
# Add other existing dependencies